phase-2/backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Validation
pydantic>=2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Gradio Interface
gradio==4.0.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
